Second Phase Submission
1. Listing1.c
It shows the execution of “ls” command using the execve() system call.
2. Listing2.c
In this we have tried to input the command from the shell itself. Also,
when we input Ctrl+c, the shell terminates. This is done   using the
system signal handlers like SIGINT, SIG_IGN.
3. Listing3.c
In this, when we enter Ctrl+c, it handles it just like a normal input
and it doesn’t terminate the shell.
4. My_Shell.c
Implementation of basic shell functions. When this code is compiled and
run it enters into our own shell and one can fire commands   like echo,
ls, cat, etc and it works similar to the original command prompt